+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/JavascriptExtension.java
@@ -138,8 +138,50 @@ public class JavascriptExtension extends FormatExtension {
 		return prettierConfig;
 	}
 
+	/**
+	 * Defaults to downloading the default Rome version from the network. To work
+	 * offline, you can specify the path to the Rome executable via
+	 * {@code rome().pathToExe(...)}.
+	 */
+	public RomeJs rome() {
+		return rome(null);
+	}
+
+	/** Downloads the given Rome version from the network. */
+	public RomeJs rome(String version) {
+		var romeConfig = new RomeJs(version);
+		addStep(romeConfig.createStep());
+		return romeConfig;
+	}
+
 	private static final String DEFAULT_PRETTIER_JS_PARSER = "babel";
 	private static final ImmutableList<String> PRETTIER_JS_PARSERS = ImmutableList.of(DEFAULT_PRETTIER_JS_PARSER, "babel-flow", "flow");
+
+	/**
+	 * Rome formatter step for JavaScript.
+	 */
+	public class RomeJs extends RomeStepConfig<RomeJs> {
+
+		/**
+		 * Creates a new Rome formatter step config for formatting JavaScript files. Unless
+		 * overwritten, the given Rome version is downloaded from the network.
+		 *
+		 * @param version Rome version to use.
+		 */
+		public RomeJs(String version) {
+			super(getProject(), JavascriptExtension.this::replaceStep, version);
+		}
+
+		@Override
+		protected String getLanguage() {
+			return "js?";
+		}
+
+		@Override
+		protected RomeJs getThis() {
+			return this;
+		}
+	}
 
 	/**
 	 * Overrides the parser to be set to a js parser.