+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/generic/Format.java
@@ -40,4 +40,8 @@ public class Format extends FormatterFactory {
 		// do not specify a default delimiter
 		return null;
 	}
+
+	public void addRome(Rome rome) {
+		addStepFactory(rome);
+	}
 }